+++ NEWS.md
@@ -9,7 +9,7 @@ border-radius: 128px;
 ## v1.9.20
 
 - Features
-    - OPDS v2 credits now link to filtered views of the credited person.
+    - OPDS v2 credits and subjects now link to views filtered on the tag.
 
 ## v1.9.19